+++ limousine_core/src/lib.rs
@@ -2,9 +2,12 @@ pub mod classical;
 pub mod component;
 pub mod index;
 pub mod iter;
+
 // pub mod learned;
 
 mod common;
+mod node_layer;
+mod traits;
 
 // Used by proc_macro
 pub use anyhow::Result;
@@ -22,6 +25,8 @@ pub use classical::BTreeTopComponent;
 
 pub use component::*;
 pub use index::*;
+pub use node_layer::*;
+pub use traits::*;
 
 pub use std::path::Path;
