+++ limousine_example/Cargo.toml
@@ -5,10 +5,7 @@ edition = "2021"
 
 [dependencies]
 average = "0.14.1"
-# eframe = "0.22.0"
-# egui = "0.22.0"
 itertools = "0.11.0"
-# btree-slab = "0.6.0"
 rand = "0.8.5"
 rand_distr = "0.4.3"
 sys-info = "0.9.1"
